+++ src/org/lukasj/idea/torquescript/editor/TSVarExpressionElementImpl.kt
@@ -1,13 +1,12 @@
-package org.lukasj.idea.torquescript.psi.impl
+package org.lukasj.idea.torquescript.editor
 
 import com.intellij.extapi.psi.ASTWrapperPsiElement
 import com.intellij.lang.ASTNode
-import com.intellij.openapi.util.NlsSafe
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReference
 import com.intellij.psi.util.elementType
-import com.intellij.util.IncorrectOperationException
+import org.lukasj.idea.torquescript.psi.TSElementFactory
 import org.lukasj.idea.torquescript.psi.TSNamedElement
 import org.lukasj.idea.torquescript.psi.TSTypes
 import org.lukasj.idea.torquescript.reference.TSGlobalVarReference
@@ -28,6 +27,6 @@ abstract class TSVarExpressionElementImpl(node: ASTNode) : ASTWrapperPsiElement(
         text
 
     override fun setName(name: String): PsiElement {
-        TODO("Not yet implemented")
+        return replace(TSElementFactory.createSimple(project, name))
     }
 }