+++ src/org/lukasj/idea/torquescript/psi/impl/TSFunctionIdentifierElementImpl.kt
@@ -5,6 +5,8 @@ import com.intellij.lang.ASTNode
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReference
+import org.lukasj.idea.torquescript.psi.TSElementFactory
+import org.lukasj.idea.torquescript.psi.TSIdentExpression
 import org.lukasj.idea.torquescript.psi.TSNamedElement
 import org.lukasj.idea.torquescript.reference.TSObjectReference
 
@@ -16,7 +18,8 @@ abstract class TSFunctionIdentifierElementImpl(node: ASTNode) : ASTWrapperPsiEle
     override fun getName(): String? = nameIdentifier.text
 
     override fun setName(name: String): PsiElement {
-        TODO("Not yet implemented")
+        nameIdentifier.replace(TSElementFactory.createSimple<TSIdentExpressionImpl>(project, name))
+        return this
     }
 
     override fun getReference(): PsiReference? =