+++ src/org/lukasj/idea/torquescript/symbols/TSGlobalVarCachedListGenerator.kt
@@ -1,17 +1,12 @@
 package org.lukasj.idea.torquescript.symbols
 
-import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiManager
 import com.intellij.psi.search.FileTypeIndex
 import com.intellij.psi.search.GlobalSearchScope
-import com.intellij.psi.util.PsiTreeUtil
-import com.intellij.psi.util.elementType
 import org.lukasj.idea.torquescript.TSFileType
-import org.lukasj.idea.torquescript.psi.TSAssignmentExpression
 import org.lukasj.idea.torquescript.psi.TSFile
-import org.lukasj.idea.torquescript.psi.TSTypes
-import org.lukasj.idea.torquescript.psi.impl.TSVarExpressionElementImpl
+import org.lukasj.idea.torquescript.editor.TSVarExpressionElementImpl
 
 class TSGlobalVarCachedListGenerator : TSCachedListGenerator<TSVarExpressionElementImpl>() {
     override fun generate(project: Project): Set<TSVarExpressionElementImpl> {